// UI/src/app/model/todoItemResponse.ts
/**
 * TodoApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Priority } from './priority';

export interface TodoItemResponse { 
    id?: number;
    name?: string;
    priority?: Priority;
    dueDate?: Date;
    description?: string;
    modifiedDate?: Date;
    isComplete?: boolean;
    groupId?: number;
}